requirements.txt
ultralytics==8.2.103
torch>=2.2.0
torchvision>=0.17.0
opencv-python-headless>=4.10.0.84
numpy>=1.26.0
gradio>=4.44.0
pyyaml>=6.0.1
tqdm>=4.66.1
pillow>=10.3.0
pytest>=7.4.0
black>=24.4.2
ruff>=0.5.0
